script/sync_datasus/pyproject.toml
[project]
name = "sync-datasus"
version = "0.1.0"
description = ""
authors = [
    {name = "Pablo Eduardo",email = "pabloeduardossantos@gmail.com"}
]
requires-python = ">=3.12,<3.14"
dependencies = [
    "pysus (>=1.0.1,<2.0.0)",
    "pandas (>=2.2.2,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pyarrow (>=23.0.1,<24.0.0)",
    "tqdm (==4.64.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
